requirements.txt
# Основные пакеты
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
alembic>=1.10.0
asyncpg>=0.27.0
pydantic>=1.10.0
python-jose>=3.3.0
cryptography>=41.0.0
passlib>=1.7.4
python-multipart>=0.0.6
httpx>=0.24.0

# Мониторинг и производительность
psutil>=5.9.0
prometheus-client>=0.17.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-prometheus>=1.18.0

# Redis и кэширование
redis>=4.5.5
aioredis>=2.0.1
hiredis>=2.2.3

# Работа с файлами и медиа
python-magic>=0.4.27
pillow>=10.0.0
aiofiles>=23.1.0

# Тестирование
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Утилиты
email-validator>=2.0.0
ruff>=0.0.290
black>=23.7.0 